move deduplicated generator to new file
